move task_details.dart to top
